CityScreenView become WeatherSceenView
